chore: update web api proj template to net6.0
