+++ src/Arcus.Templates.Tests.Integration/Fixture/TargetFramework.cs
@@ -8,6 +8,11 @@
         /// <summary>
         /// Sets the target framework of the project created from the templates to netcoreapp3.1.
         /// </summary>
-        NetCoreApp31
+        NetCoreApp31,
+
+        /// <summary>
+        /// Sets the target framework of the project created from the templates to net6.0.
+        /// </summary>
+        Net6_0
     }
 }
