+++ src/Arcus.Templates.Tests.Integration/WebApi/Health/v1/HealthEndpointTests.cs
@@ -3,7 +3,9 @@ using System.Net.Http;
 using System.Threading.Tasks;
 using Arcus.Templates.Tests.Integration.Fixture;
 using Microsoft.Extensions.Diagnostics.HealthChecks;
+
 using Newtonsoft.Json;
+
 using Xunit;
 using Xunit.Abstractions;
 
@@ -13,12 +15,10 @@ namespace Arcus.Templates.Tests.Integration.WebApi.Health.v1
     [Trait("Category", TestTraits.Integration)]
     public class HealthEndpointTests
     {
-        private readonly TestConfig _configuration;
         private readonly ITestOutputHelper _outputWriter;
 
         public HealthEndpointTests(ITestOutputHelper outputWriter)
         {
-            _configuration = TestConfig.Create();
             _outputWriter = outputWriter;
         }
 
@@ -26,18 +26,20 @@ namespace Arcus.Templates.Tests.Integration.WebApi.Health.v1
         public async Task Health_Get_Succeeds()
         {
             // Arrange
-            using (WebApiProject project = await WebApiProject.StartNewAsync(_configuration, _outputWriter))
-                // Act
-            using (HttpResponseMessage response = await project.Health.GetAsync())
+            using (WebApiProject project = await WebApiProject.StartNewAsync(_outputWriter))
             {
-                // Assert
-                Assert.Equal(HttpStatusCode.OK, response.StatusCode);
+                // Act
+                using (HttpResponseMessage response = await project.Health.GetAsync())
+                {
+                    // Assert
+                    Assert.Equal(HttpStatusCode.OK, response.StatusCode);
 
-                string healthReportJson = await response.Content.ReadAsStringAsync();
+                    string healthReportJson = await response.Content.ReadAsStringAsync();
 
-                var healthReport = JsonConvert.DeserializeObject<HealthReport>(healthReportJson, new TimeSpanConverter());
-                Assert.NotNull(healthReport);
-                Assert.Equal(HealthStatus.Healthy, healthReport.Status);
+                    var healthReport = JsonConvert.DeserializeObject<HealthReport>(healthReportJson);
+                    Assert.NotNull(healthReport);
+                    Assert.Equal(HealthStatus.Healthy, healthReport.Status);
+                }
             }
         }
     }
